forkorting av importer
