+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Routers from "./core/routes/Routers";
-import { Navbar } from "@components/Navbar/Navbar";
+import { Navbar } from "@components/Navbar";
 import { useAuthContext } from "@auth/useAuthContext";
 import { Toaster } from "react-hot-toast";
 import "@/index.css";
