+++ client/src/components/OverviewPage/Map/SeatPaths.ts
