+++ client/src/components/OverviewPage/SeatInfo/index.tsx
