+++ client/src/components/OverviewPage/index.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from "react";
-import Heading from "@common-components/Heading";
+import { Heading } from "@common-components/Heading";
 import { useAuthContext } from "@auth/useAuthContext";
 import { observer } from "mobx-react-lite";
-import OverviewMap from "@components/OverviewPage/OverviewMap/OverviewMap";
-import SeatInfo from "@components/OverviewPage/OverviewSeatInfo/OverviewSeatInfo";
-import DateSwitchButton from "@components/OverviewPage/OverviewDateSwitchButton/OverviewDateSwitchButton";
+import OverviewMap from "@components/OverviewPage/Map";
+import SeatInfo from "@components/OverviewPage/SeatInfo";
+import { DateSwitchButton } from "@components/OverviewPage/DateSwitchButton";
 import bookingStore from "@stores/BookingStore";
-import OverviewEventModeButton from "./OverviewEventModeButton/OverviewEventModeButton";
+import { EventModeButton } from "./EventModeButton";
 import { Calendar, DateObject } from "react-multi-date-picker";
 
 const OverviewPage = observer(() => {
@@ -47,7 +47,7 @@ const OverviewPage = observer(() => {
       <div className="flex flex-col justify-center items-center h-full">
         <div className="flex flex-col gap-10">
           <Heading title={welcomeTitle} subTitle={subTitle} />
-          {isEventAdmin ? <OverviewEventModeButton /> : <DateSwitchButton />}
+          {isEventAdmin ? <EventModeButton /> : <DateSwitchButton />}
           <div className="flex justify-center">
             {bookingStore.bookEventMode &&
             bookingStore.isEventDateChosen === false ? (
