+++ client/src/core/auth/IsAuthenticated.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import { useAuthContext } from "@auth/useAuthContext";
 import { useEffect } from "react";
-import Loading from "@common-components/Loading";
+import {Loading} from "@common-components/Loading";
 
 export const IsAuthenticated = () => {
   const navigate = useNavigate();
